+++ README.md
@@ -17,7 +17,7 @@ dependencies {
 2) Add the following permission to your `AndroidManifest.xml`
 
 ```xml
-<uses-permission android:name="com.owncloud.android.sso"/>
+<uses-permission android:name="com.nextcloud.android.sso"/>
 ```
 
 3) To choose an account, include the following code in your login dialog:
