+++ build.gradle
@@ -22,7 +22,6 @@ android {
     compileSdkVersion 27
     buildToolsVersion '27.0.3'
     defaultConfig {
-        //applicationId 'de.luhmer.owncloud.accountimporter'
         minSdkVersion 14
         targetSdkVersion 27
         versionCode 1
