+++ src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="de.luhmer.owncloud.accountimporter">
+    package="com.nextcloud.android.sso">
 
-    <uses-permission android:name="com.owncloud.android.sso" />
+    <uses-permission android:name="com.nextcloud.android.sso" />
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <uses-permission android:name="android.permission.GET_ACCOUNTS" />
