--- src/main/java/de/luhmer/owncloud/accountimporter/aidl/NextcloudRequest.java
@@ -1,87 +1,0 @@
-package de.luhmer.owncloud.accountimporter.aidl;
-
-import java.io.Serializable;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- *  Nextcloud SingleSignOn
- *
- *  @author David Luhmer
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-public class NextcloudRequest implements Serializable {
-
-    static final long serialVersionUID = 215521212534236L; //assign a long value
-
-    public String method;
-    public Map<String, List<String>> header = new HashMap<>();
-    public Map<String, String> parameter = new HashMap<>();
-    public String requestBody;
-    public String url;
-    public String token;
-    public String accountName;
-
-    private NextcloudRequest() { }
-
-    public static class Builder {
-        private NextcloudRequest ncr;
-
-        public Builder() {
-            ncr = new NextcloudRequest();
-        }
-
-        public NextcloudRequest build() {
-            return ncr;
-        }
-
-        public Builder setMethod(String method) {
-            ncr.method = method;
-            return this;
-        }
-
-        public Builder setHeader(Map<String, List<String>> header) {
-            ncr.header = header;
-            return this;
-        }
-
-        public Builder setParameter(HashMap<String, String> parameter) {
-            ncr.parameter = parameter;
-            return this;
-        }
-
-        public Builder setRequestBody(String requestBody) {
-            ncr.requestBody = requestBody;
-            return this;
-        }
-
-        public Builder setUrl(String url) {
-            ncr.url = url;
-            return this;
-        }
-
-        public Builder setToken(String token) {
-            ncr.token = token;
-            return this;
-        }
-
-        public Builder setAccountName(String accountName) {
-            ncr.accountName = accountName;
-            return this;
-        }
-    }
-}
